archive/samples/python/24.bot-authentication-msgraph/requirements.txt
botbuilder-integration-aiohttp>=4.14.0
botbuilder-dialogs>=4.14.0
requests_oauthlib

